Revert "Fixes various thruster-related bugs" (#4779)
## About The Pull Request

Reverts shiptest-ss13/Shiptest#4670

## Why It's Good For The Game

All the fixes slowed pretty much every ship to a crawl, even
subshuttles. I tried adjusting thrust on all engines to compensate for
the pr but it was extremely weird to set everything so high and even
weirder with combustions, where they get to pill-speeds at even just 24
because of how they react to different gas mixes and parts.

## Changelog

:cl:
del: Reverts fixes to plasma and ion thrusters that ended up slowing
down ships to a crawl
/:cl: