Fixes and Improves Turret Controls (#4751)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Ports: https://github.com/PentestSS13/Pentest/pull/279

Turret controls could not be removed (or attacked with items) once
placed, this corrects that. You are now able to attack them with stuff
or dismantle them with a wrench if the controls are unlocked. Also fixes
issues with mounting them, as they were mounted behind you.

Changed the way turret controls are linked. They are now linked from the
main turret control to the turret, so you don't have to run around as
much linking them together. Previously they were linked from the turret
to the controls, so you needed to get each turret's buffer.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Fix and improvement. Hopefully this will make an easier time for those
who like to build their own turret arrays.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Turret controls can be taken down with a crowbar and face the right
direction when mounted on a wall.
refactor: Turret controls are now linked to individual turrets through
their buffer instead of the other way around (minimizes running back and
forth).
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FslashN <150570209+FslashN@users.noreply.github.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>