Fixes fax spam only sending a single piece of mail (#4802)
## About The Pull Request

Calling receive() on a fax physically moves the actual paper. The issue
is the event only initializes a single paper, so it'd end up juggling it
around inbetween faxes.

So we need to initialize a new one for each fax we target with our spam.


![image](https://github.com/user-attachments/assets/e911af9c-321a-47c9-8ed0-142d24ce0b8a)

## Why It's Good For The Game

Bug fix good

closes #4770

## Changelog

:cl:
fix: The spam fax event now actually spams faxes
/:cl: