Icemoon Training Center Adjustments (#4732)
## About The Pull Request

Slightly adjusts the mech arena and the mech's loadout itself to be less
cheesable.

![image](https://github.com/user-attachments/assets/e1f26b7f-aa85-489a-93fc-1aa521200f80)

Removes the mech's LMG as it does not function right with the pilot's
AI, and adds an additional armor plate.

## Why It's Good For The Game

People should not be cheesing the gygax boss for a free dark gygax.

## Changelog

:cl:
balance: Icemoon Training Center mech is no longer as easily cheesable.
/:cl: