Sunskipper grill fix (#4812)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
#4739 removed grill coal from the sunskipper, as well as the ability to
use coal in grills at all. This _**TEMPORARILY**_ replaces it with wood.
I'll revert this when grills get un-nerfed.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Did you know that coal was a material before OM5, used exclusively for
grills? That was dumb, but wood sucks ass and I'm very sad that I
**have** to use it now.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: You can use the Sunskipper grill again.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Theos <theubernyan@gmail.com>