Removes access restrictions on Outpost Maint airlocks (#4848)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Removes the access restrictions on most of the outpost maint airlocks on
indie space and NT ice. Doors that lead to restricted areas are still
locked.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Discussed during maptainer meet a while ago and I've been meaning to get
around to doing this. The outpost maints in general aren't actually
restricted areas, and the locks on them created the implication that
they in fact were. Having them open should provide more interesting
avenues for getting around the outpost as well.

## Changelog

:cl:
del: Outpost maint access restrictions.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->