Scrapstation Touch Ups (#4847)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Removes extra APC terminals on the wreck

Foreman's 357 is now a Coalition version rather than indie

Fixes armory shutter IDs

Removes a boomslang guy near the wreck and replaces one of the heavy
turrets on the solar array with a strike turret.

## Why It's Good For The Game

Fixes are good. The abundance of long range sniper enemies was also a
bit oppressive in a zero-g environment, so I culled a few of them.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Scrapping station wreck no longer has extra terminals, foreman's
office has correct revolver, armory shutters IDs
balance: Removes a deadeye and replaces a revolt turret with a strike on
Scrapping Station
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->