Medical Aid kit and Advanced First Aid kit comes with painkillers (#4854)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds a pill bottle of tramal to the medical aid kit and a bottle of
dimorlin and a syringe to the advanced first aid kit.

Increases the max items on the advanced aid kit to 10 so it can fit
everything.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Since we've moved to painkiller gameplay for surgery, it makes sense to
have painkillers in the relevant medical kits. Especially in the case of
the medical aid kit, having a starting source of painkillers for surgery
should help normalize and encourage their use in gameplay.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Medical Aid kit and Advanced First Aid kit now have painkillers.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->