Makes it easier to poke and annoy your colleagues than ever before. Also makes it so that hugging or shaking someone's hand requires you to grab them first. (#4868)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR changes hugging so that it requires that one grabs the person
they wish to hug first. If one clicks on another character without
grabbing someone, they end up poking them instead. This PR also makes it
so that you need to grab someone in order to shake their hand.
Otherwise, you just pat the person on the shoulder instead. This PR also
fixes a bug where shaking someone up from a lying down position would
call both text for shaking them up, as well as text for a specific body
region.

Before:

![image](https://github.com/user-attachments/assets/6243bd6d-a107-4a31-9ed0-d882068d9510)

After:

![image](https://github.com/user-attachments/assets/e84425eb-b8f4-4309-8cb3-599076718b85)


## Why It's Good For The Game

I've always found it a little silly that one could accidentally click on
another character and hug them that way. This change should make it so
that the more realistic scenario of tapping or poking someone happens
instead. It also ensures that if someone wants to hug someone else,
there is proper intent behind it. This also ties into bear hugs, which
require an aggressive grab to function, so this would end up making hugs
follow a logical sequence of intent.

It also makes sense to have to grab someone before shaking their hand, I
think, for a similar reason, in that it signals intent better.

## Changelog

:cl:
balance: hugs now require grabbing the person to be hugged -- not doing
so pokes that person instead. The same is true for shaking one's hand.
fix: fixes shaking someone up also poking them or patting them at the
same time
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->