Thruster speed on helm consoles is now accurate (#4817)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

based on #4670 except without the part that changes thruster output
In the event the physics subsystem has its speed modified this should
allow that to be handled in real time by the helm console on update

## Why It's Good For The Game

Tells you how fast you can ACTUALLY expect a ship to speed up in seconds
rather than not doing that

## Changelog

:cl: Sapphicoverload, theos
tweak: Thruster speed on helm consoles is now accurate to how fast the
ship accelerates
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->