Fixes om5 smooth tile projectile offset (#4865)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Accounts for the offset in certain turfs causing projectiles to fly off
at a weird angle
There is probably a better way of doing this but this one works

## Why It's Good For The Game

I personally like it when I click somewhere to shoot something and the
gun I'm shooting shoots there instead of somewhere else

## Changelog

:cl:
fix: Guns will no longer decide to fire incorrectly sometimes
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->